// src/app/Models/phone.model.ts
export class Phone {
    BusinnesID: number;
    CustomerID: number;
    ID: number;
    PhoneNumber: string;
    PhoneCode: string;
    Annexed: string;
    Class: string;
    ClassName: string;
    Origin: string;
    OriginName: string;
    Provider: string;
    ProviderName: string;
    Priority: number;
    SubPriority: number;
    Observation: string;
    State: string;
}